Fixed double to uint32_t cast in getTimestampFromSeconds()

+++ src/rtppacketizationconfig.cpp
@@ -64,7 +64,7 @@ double RtpPacketizationConfig::timestampToSeconds(uint32_t timestamp) {
 }
 
 uint32_t RtpPacketizationConfig::getTimestampFromSeconds(double seconds, uint32_t clockRate) {
-	return uint32_t(seconds * clockRate);
+	return uint32_t(int64_t(seconds * double(clockRate))); // convert to integer then cast to u32
 }
 
 uint32_t RtpPacketizationConfig::secondsToTimestamp(double seconds) {
